+++ tests/utils.py
@@ -7,4 +7,4 @@ from validator.utils import get_env_array
 class TestUtils(TestCase):
     def test_getEnvArray(self):
         os.environ['INPUT_ONLY'] = '["fileA.version", "fileB.txt"]'
-        self.assertEquals(get_env_array('INPUT_ONLY'), ['fileA.version', 'fileB.txt'])
+        self.assertEqual(get_env_array('INPUT_ONLY'), ['fileA.version', 'fileB.txt'])